+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod srt;
 
 pub mod rtmp;
 pub mod rtsp;
+pub mod s3;
 pub mod sora;
 pub(crate) mod sora_publisher;
 pub mod stats;